requirement.txt
Appium-Python-Client
selenium
pytest